Move components to a folder
