+++ src/components/properties.js
